pss, trojan: create trojan package, move trojan code to trojan package, rename trojan_chunk.go to message.go and likewise with test
