+++ pss/trojan/message.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Swarm library. If not, see <http://www.gnu.org/licenses/>.
 
-package pss
+package trojan
 
 import (
 	"bytes"
